+++ TestProjects/AndroidBigPicture/Assets/Plugins/Android/FileAccess.androidlib/src/main/java/com/unity/fileaccess/ImageProvider.java
@@ -1,0 +1,10 @@
+package com.unity.fileaccess;
+
+import androidx.core.content.FileProvider;
+import com.unity.fileaccess.R;
+
+public class ImageProvider extends FileProvider {
+    public ImageProvider() {
+        super(R.xml.file_paths);
+    }
+}